pyproject.toml
[project]
name = "quant-binance-spot"
version = "0.1.0"
requires-python = ">=3.10,<3.13"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "pyyaml>=6.0",
  "python-dotenv>=1.0",
  "requests>=2.31",
  "tqdm>=4.66",
  "vectorbt>=0.26.0",
  "numba==0.58.1",
  "llvmlite==0.41.1",
  "pyarrow>=14.0.0",
  "matplotlib>=3.7.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
